Fix infloop in password-store--run
For certain commands (like password-store--run-generate,
password-store--run-remove) with password entries in sub-directories,
password-store--run would infloop. The output variable never received any
actual output. This seems to be a timing issue as e.g. stepping through the
code in edebug never caused the issue.

This commit changes the check for the finished pass process and also inserts
calls to accept-process-output in order to guarantee that output really gets
processed while the loop is running.